--- configuration/jardb.properties
@@ -1,1 +1,0 @@
-replication.meta_servers=10.235.114.5:34602,10.235.114.5:34601,10.235.114.5:34603
