+++ configuration/pegasus.properties
@@ -1,0 +1,1 @@
+meta_servers=127.0.0.1:34601,127.0.0.1:34602,127.0.0.1:34603
