+++ scripts/run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+java -classpath ${project.build.finalName}.jar:library/* com.xiaomi.infra.pegasus.client.$*
+
